separar codigo de mock segun el ensayo
